+++ lib/widgets/editplus_quickform.dart
@@ -2,12 +2,12 @@ part of edit_plus;
 
 class EditplusQuickform extends StatelessWidget
 {
-  final List<EditPlusFormWidget> widgetDescList;
-  final Map savedValuesMap;
+  final List widgetDescList;
+  final Map dataContainer;
   final GlobalKey formKey;
   final double spacing;
 
-  EditplusQuickform({this.widgetDescList, this.savedValuesMap, this.formKey, this.spacing});
+  EditplusQuickform({this.widgetDescList, this.dataContainer, this.formKey, this.spacing});
 
   Widget build(BuildContext context) {
     return Form(
@@ -22,28 +22,75 @@ class EditplusQuickform extends StatelessWidget
   {
     var widgetList = List<Widget>();
 
-    for (EditPlusFormWidget entry in widgetDescList)
+    for (var entry in widgetDescList)
     {
-      if (entry.widgettype == 'TEXTFIELD')
+      if (entry is EditPlusFormWidget)
       {
-        var tfwidget = TextFormField(
-          decoration: InputDecoration(
-            labelText:  entry.label,
-            border: OutlineInputBorder(),
-          ),
-          onSaved: (String value) {
-            savedValuesMap[entry.savekey] = value;
-          },
-          validator: entry.validationfunction,
-        );
-        widgetList.add(tfwidget);
-        widgetList.add(SizedBox(height: spacing,));
-        continue;
+        if (entry.widgettype == 'TEXTFIELD')
+        {
+          var tfwidget = TextFormField(
+            decoration: InputDecoration(
+              labelText:  entry.label,
+              border: OutlineInputBorder(),
+            ),
+            onSaved: (String value) {
+              dataContainer[entry.savekey] = value;
+            },
+            validator: entry.validationfunction,
+            onChanged: entry.onchangefunction,
+          );
+          widgetList.add(tfwidget);
+          widgetList.add(SizedBox(height: spacing,));
+          continue;
+        }
+        if (entry.widgettype == 'PASSWORDINPUTFIELD')
+        {
+          var tfwidget = TextFormField(
+            decoration: InputDecoration(
+              labelText:  entry.label,
+              border: OutlineInputBorder(),
+            ),
+            onSaved: (String value) {
+              onSavedFunction(entry.savekey, value);
+            },
+            validator: entry.validationfunction,
+            onChanged: entry.onchangefunction,
+            obscureText: true,
+            obscuringCharacter: "*",
+          );
+          widgetList.add(tfwidget);
+          widgetList.add(SizedBox(height: spacing,));
+          continue;
+        }
+        if (entry.widgettype == 'DATEINPUTFIELD')
+        {
+          var tfwidget = EditPlusDateInputFormField(
+              controller: new TextEditingController(),
+              validationFunction: entry.validationfunction,
+              label: entry.label,
+              saveDataKey: entry.savekey,
+              onSaveFunction: onSavedFunction,
+              // formDataContainer: dataContainer,
+            );
+
+          widgetList.add(tfwidget);
+          widgetList.add(SizedBox(height: spacing,));
+          continue;
+        }        
+      }
+      if (entry is Widget)
+      {
+        widgetList.add(entry);
       }
 
     }
     
     return widgetList;
+  }
+
+  onSavedFunction(var key, var value)
+  {
+    dataContainer[key] = value;
   }
 
 }
@@ -54,6 +101,9 @@ class EditPlusFormWidget
   final String label;
   final String savekey;
   final Function validationfunction;
+  final Function onchangefunction;
+  final int size; /// Valid for text inputs which have a maximum size
+  final otherData;
 
-  EditPlusFormWidget({this.widgettype, this.label, this.savekey, this.validationfunction});
+  EditPlusFormWidget({this.widgettype, this.label, this.savekey, this.validationfunction, this.onchangefunction, this.size, this.otherData});
 }